+++ src/main/java/org/jboss/vfs/spi/JavaZipFileSystem.java
@@ -165,8 +165,11 @@ public final class JavaZipFileSystem implements FileSystem {
         return zipFile.getInputStream(entry);
     }
 
-    public boolean delete(VirtualFile mountPoint, VirtualFile target) throws IOException {
-        final ZipNode zipNode = getExistingZipNode(mountPoint, target);
+    public boolean delete(VirtualFile mountPoint, VirtualFile target) {
+        final ZipNode zipNode = getZipNode(mountPoint, target);
+        if (zipNode == null) {
+            return false;
+        }
         final File cachedFile = zipNode.cachedFile;
         return cachedFile != null && cachedFile.delete();
     }
@@ -188,7 +191,7 @@ public final class JavaZipFileSystem implements FileSystem {
         return cachedFile != null ? cachedFile.lastModified() : entry == null ? zipTime : entry.getTime();
     }
 
-    public boolean exists(VirtualFile mountPoint, VirtualFile target) throws IOException {
+    public boolean exists(VirtualFile mountPoint, VirtualFile target) {
         final ZipNode zipNode = rootNode.find(mountPoint, target);
         if (zipNode == null) {
             return false;
@@ -203,8 +206,11 @@ public final class JavaZipFileSystem implements FileSystem {
         return zipNode != null && zipNode.entry == null;
     }
 
-    public List<String> getDirectoryEntries(VirtualFile mountPoint, VirtualFile target) throws IOException {
-        final ZipNode zipNode = getExistingZipNode(mountPoint, target);
+    public List<String> getDirectoryEntries(VirtualFile mountPoint, VirtualFile target) {
+        final ZipNode zipNode = getZipNode(mountPoint, target);
+        if (zipNode == null) {
+            return Collections.emptyList();
+        }
         final Map<String, ZipNode> children = zipNode.children;
         if (children == null) {
             return Collections.emptyList();
@@ -224,6 +230,14 @@ public final class JavaZipFileSystem implements FileSystem {
             throw new IOException("Cannot call this operation on a directory");
         }
         return entry;
+    }
+
+    private ZipNode getZipNode(VirtualFile mountPoint, VirtualFile target) {
+        final ZipNode zipNode = rootNode.find(mountPoint, target);
+        if (zipNode == null) {
+            return null;
+        }
+        return zipNode;
     }
 
     private ZipNode getExistingZipNode(VirtualFile mountPoint, VirtualFile target)
